ingestion/requirements.txt
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.0
google-cloud-bigquery-storage==2.25.0
pyarrow==15.0.2
pandas==2.2.2
requests==2.31.0
tenacity==8.3.0
structlog==24.2.0
click==8.1.7
